Conclude Japanese localization for the login view
Japanese characters behave differently than western languages,
in that they also use full-width characters. Therefore, if one
wants to move a button horizontally based on its text length,
consider using an offset of -4 (value determined experimentally)
for the button styles (i.e. brackets and spaces). This gets more
complicated if its default attribute is turned on, but since I
am not using it anywhere we don't have to worry about that for
now.

Maybe I will add a IsAsian property in the future should I decide
to add Mandarin language support, though this is a story for
another time.